Initial Commit :
HarkonReddit is a web app used to monitor reddit posts from hot subs. This is a purely fun, half vibe coded project to see the affect of public opinion on stock performance. From now, commits will be more strict and monitored. The project will be made into a complete SDK to explore and experiment for various geeksheads and also will be used to test any relevance to actual stock performances. The sentimental analysis will be improved in future versions.